requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.3=py35h28b3542_0
asn1crypto=0.24.0=py35_0
attrs=18.2.0=py35h28b3542_0
automat=0.7.0=py35_0
backcall=0.1.0=py35_0
bidict=0.18.0=pypi_0
blas=1.0=mkl
bleach=2.1.4=py35_0
blosc=1.7.0=vc14_2
bokeh=0.13.0=py35_0
boost=1.59.0=py35_0
boost-cpp=1.68.0=h6a4c333_1000
bzip2=1.0.6=vc14_3
ca-certificates=2019.10.16=0
certifi=2018.8.24=py35_1
cffi=1.11.5=py35h74b6da3_1
chardet=3.0.4=py35_1
cloudpickle=0.7.0=py_0
colorama=0.3.9=py35h32a752f_0
constantly=15.1.0=py35h28b3542_0
cryptography=1.0.2=pypi_0
cuda92=1.0=0
cudatoolkit=9.0=1
curl=7.65.3=h2a8f88b_0
cycler=0.10.0=py35hcc71164_0
dask=0.15.2=py35_0
dask-core=1.1.5=py_0
decorator=4.3.0=py35_0
dlib=18.18=py35_2
entrypoints=0.2.3=py35_2
expat=2.2.5=he025d50_0
freetype=2.8=vc14h17c9bdf_0
freexl=1.0.5=hfa6e2cd_0
gdal=2.2.2=py35h202a4aa_1
geos=3.6.2=h9ef7328_2
h5py=2.9.0=pypi_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.1=vc14hb361328_0
html5lib=1.0.1=py35_0
hyperlink=17.3.1=py_0
icc_rt=2017.0.4=h97af966_0
icu=58.2=vc14hc45fdbb_0
idna=2.0=py35_0
imageio=2.4.1=py35_0
imbalanced-learn=0.4.3=py_0
incremental=17.5.0=py35_0
intel-openmp=2019.1=144
ipykernel=4.9.0=py35_1
ipython=6.5.0=py35_0
ipython-notebook=4.0.4=py35_3
ipython_genutils=0.2.0=py35ha709e79_0
jedi=0.12.1=py35_0
jinja2=2.10=py35_0
jpeg=9b=vc14h4d7706e_1
jsonschema=2.6.0=py35h27d56d3_0
jupyter_client=5.2.3=py35_0
jupyter_core=4.4.0=py35_0
kealib=1.4.7=ha5b336b_5
kiwisolver=1.0.1=py35h6538335_0
krb5=1.16.1=hc04afaa_7
libboost=1.65.1=vc14h7b94727_3
libcurl=7.65.3=h2a8f88b_0
libflang=5.0.0=h6538335_20180525
libgdal=2.2.2=h2727f2b_1
libiconv=1.15=h1df5818_7
libkml=1.3.0=hc65d273_3
libnetcdf=4.4.1.1=h825a56a_8
libopenblas=0.3.7=h29e5d5d_1
libpng=1.6.37=h7602738_0
libpq=11.2=h3235a2c_0
libprotobuf=3.2.0=vc14_0
libsodium=1.0.3=0
libspatialite=4.3.0a=h383548d_18
libssh2=1.8.2=h7a1dbc1_0
libtiff=4.0.9=h36446d0_2
libxml2=2.9.9=h464c3ec_0
llvm-meta=5.0.0=0
locket=0.2.0=py35h0dfcdd0_1
lz4-c=1.8.1.2=h2fa13f4_0
lzo=2.10=vc14h0a64fa6_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.0=py35hfa6e2cd_1
matplotlib=2.2.2=py35h153e9ff_1
mistune=0.8.3=py35hfa6e2cd_1
mkl=2018.0.3=1
mkl_fft=1.0.9=py35_0
mkl_random=1.0.1=py35_0
msys2-conda-epoch=20160418=1
nbconvert=5.3.1=py35_0
nbformat=4.4.0=py35h908c9d9_0
networkx=2.1=py35_0
ninja=1.8.2=py35he980bc4_1
notebook=5.6.0=py35_0
numexpr=2.6.8=py35h9ef55f4_0
numpy=1.15.2=py35ha559c80_0
numpy-base=1.15.2=py35h8128ebf_0
olefile=0.46=py35_0
openblas=0.3.7=h828a276_1
opencv3=3.1.0=py35_0
openjpeg=2.3.0=h5ec785f_1
openmp=5.0.0=vc14_1
openssl=1.1.1d=he774522_3
packaging=17.1=py35_0
pandas=0.23.4=py35h830ac7b_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py35_1
parso=0.3.1=py35_0
partd=0.3.10=py_0
patsy=0.5.0=py35_0
pickleshare=0.7.4=py35h2f9f535_0
pillow=5.2.0=pypi_0
pip=10.0.1=py35_0
plotly=3.10.0=py_0
plotly-orca=1.2.1=1
proj4=4.9.3=hcf24537_7
prometheus_client=0.3.1=py35h28b3542_0
prompt_toolkit=1.0.15=py35h89c7cb4_0
psutil=5.4.7=py35hfa6e2cd_0
pyasn1=0.1.9=py35_0
pyasn1-modules=0.0.9=py35_0
pycparser=2.18=py35_1
pygments=2.2.0=py35h24c0941_0
pylsl=1.13.1=pypi_0
pyopenssl=0.15.1=py35_2
pyparsing=2.2.0=py35_1
pyqt=5.9.2=py35h6538335_2
pyshp=2.1.0=py_0
pysocks=1.6.8=py35_0
pystasm=0.3.1=pypi_0
pytables=3.4.2=py35h9e98cc6_2
python=3.5.6=he025d50_0
python-dateutil=2.7.3=py35_0
pytorch=1.0.1=py3.5_cuda90_cudnn7_1
pytz=2018.5=py35_0
pywavelets=1.0.0=py35h452e1ab_0
pywin32=223=py35hfa6e2cd_1
pywinpty=0.5.4=py35_0
pyyaml=3.13=py35hfa6e2cd_0
pyzmq=17.0.0=py35hfa6e2cd_0
qt=5.9.7=vc14h73c81de_0
requests=2.13.0=py35_0
retrying=1.3.3=py35_2
scikit-image=0.13.1=py35hfa6e2cd_1
scikit-learn=0.20.0=py35heebcf9a_1
scipy=1.1.0=py35h4f6bf74_1
seaborn=0.9.0=py35_0
send2trash=1.5.0=py35_0
service_identity=17.0.0=py35h28b3542_0
setuptools=40.2.0=py35_0
shapely=1.6.4=py35hc90234e_0
shortuuid=0.5.0=pypi_0
simplegeneric=0.8.1=py35_2
sip=4.19.8=py35h6538335_0
six=1.11.0=pypi_0
snappy=1.1.7=vc14h2dea872_1
sqlite=3.29.0=hfa6e2cd_1
statsmodels=0.9.0=py35h452e1ab_0
tbb=2019.4=h74a9793_0
tbb4py=2018.0.5=py35he980bc4_0
terminado=0.8.1=py35_1
testpath=0.3.1=py35h06cf69e_0
thrift-cpp=0.12.0=h59828bf_1002
tk=8.6.8=hfa6e2cd_0
toolz=0.9.0=py35_0
torchvision=0.2.1=pypi_0
tornado=5.1.1=py35hfa6e2cd_0
tqdm=4.36.1=py_0
traitlets=4.3.2=py35h09b975b_0
twisted=17.5.0=py35_0
urllib3=1.20=py35_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wcwidth=0.1.7=py35h6e80d8a_0
webencodings=0.5.1=py35_1
wheel=0.31.1=py35_0
win_inet_pton=1.0.1=py35_1
win_unicode_console=0.5=py35h56988b5_0
wincertstore=0.2=py35hfebbdb8_0
winpty=0.4.3=4
xerces-c=3.2.2=ha925a31_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=vc14h4cb57cf_1
zeromq=4.1.3=vc14_1
zlib=1.2.11=vc14h1cdd9ab_1
zope=1.0=py35_1
zope.interface=4.5.0=py35hfa6e2cd_0
zstd=1.3.3=vc14_1
